refactor(mobile): Push notification subscription (#5924)
* feat(store): add hooks into the cgwClient lifecycle

The Auth endpoint returns a response cookie. This cookie is automatically set by browsers, but this doesn’t work that way in React-Native context. Because of this we need to be able to intercept the requests and deal with the cookie ourselves. By providing those hooks (or „escape hatches“), we can hook into the request process and manage the hooks on the mobile side.

* feat(mobile): handle response cookie

* refactor: move delegate typedData struct to utils

* feat(mobile): register delegate after key import

* fix(mobile): fix crash when loading accounts data

* fix(mobile): register for queue notifications

We were not properly authenticating the user and this was leading to the user not being subscribed for push notifications for queue events

* feat(mobile): use middleware for push notifications

* refactor(mobile): simplify notification code

* feat(mobile): create per safe notification subscriptions

* fix(mobile): update account when adding a delegate

* fix(mobile): linting and failing tests

* feat(mobile): add global notification setting

* fix(mobile): subscription to notifications is handled in middleware

* fix(mobile): only update subscription if safe was subscribed for notifications

* refactor(mobile): useGTW hook and registration should use same code

* refactor(mobile): requestAndRegister helper

Created a new requestAndRegister helper within useNotificationManager to consolidate permission checks and registration logic

Reused this helper when enabling notifications, simplifying the permission prompt flow

Applied the helper in toggleNotificationState to remove duplicated permission logic

* refactor(mobile): getDelegateKeyId

Added a reusable getDelegateKeyId helper in the mobile utilities package, providing a single location to construct delegate key IDs

* refactor(mobile): getAccountType

introduced a reusable helper to derive notification account type based on Safe owners and signers

* fix(mobile): cleanup console.logs and comments

* fix(mobile): linting

* fix(mobile): wrong comment

* fix(mobile): resolve rebase issues

* fix(mobile): require cycle

* fix(mobile): failing test

* fix(mobile): remove useMyAccounts hook

It was already removed but the branch rebase was not correct.

* feat(mobile): add notification parser

* fix(mobile): wrong variable used

* feat(mobile): display push notifications on foreground

* fix(mobile): remove useGTW

* fix(mobile): go around circular dependencies issue

* fix(mobile): failing tests

* refactor(mobile): use strategy pattern in middleware